requirements.txt
openai
python-pptx
python-dotenv